Move `backend.python.lint.mypy` to `backend.python.typecheck.mypy`
This is more consistent, e.g. making docs easier to approach. Also, there are multiple Python type checkers, so this accomadates any possible future Python type checkers we may add.

# Rust tests will be skipped. Delete if not intended.
[ci skip-rust-tests]
