+++ pants.toml
@@ -31,7 +31,7 @@ backend_packages2.add = [
   "pants.backend.python.lint.docformatter",
   "pants.backend.python.lint.flake8",
   "pants.backend.python.lint.isort",
-  "pants.backend.python.lint.mypy",
+  "pants.backend.python.typecheck.mypy",
   "internal_backend.rules_for_testing",
   "internal_backend.utilities",
 ]
