+++ src/python/pants/init/BUILD
@@ -58,8 +58,8 @@ target(
     'src/python/pants/backend/python/lint/docformatter',
     'src/python/pants/backend/python/lint/flake8',
     'src/python/pants/backend/python/lint/isort',
-    'src/python/pants/backend/python/lint/mypy',
     'src/python/pants/backend/python/lint/pylint',
+    'src/python/pants/backend/python/typecheck/mypy',
     'src/python/pants/backend/native',
     'src/python/pants/core:plugin',
   ],
